fix: wrong markdown syntax highlighting
