+++ apps/editor/src/css/md-syntax-highlighting.css
@@ -75,7 +75,7 @@
   color: #555;
 }
 
-.tui-md-table .tui-md-marked-text {
+.tui-md-table .tui-md-table-cell {
   color: #222;
 }
